+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # MForm - REDAXO Addon für Modul-Input-Formulare
 
-## Unveröffentlicht (9.x)
+## Version 9.5.1
 
 ### Behoben